requirements.txt
# LLM Providers
litellm>=1.0.0
openai>=1.0.0
anthropic>=0.25.0

# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0

# Async & Concurrency
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
